sentence-ranking/requirements.txt
backoff==2.2.1
fastapi~=0.82.0
openai~=0.27.8
python-dotenv==1.0.0
tiktoken==0.4.0
uvicorn~=0.18.3